magellon-sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/magellon_sdk"]

[project]
name = "magellon-sdk"
version = "1.2.0"
description = "SDK for building Magellon plugins: PluginBase, ProgressReporter, CloudEvents envelope, and executor contracts."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Behdad Khoshbin", email = "b.khoshbin@gmail.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "pydantic>=2.0",
    # Required by magellon_sdk.archive.manifest (H3a plugin archive format)
    # and by the `magellon-sdk plugin` CLI. Every plugin ships with this
    # already (via the [config] extra) so promoting it to core is free.
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
]
rabbitmq = [
    "pika>=1.3",
]
nats = [
    "nats-py>=2.6",
]
logging = [
    "python-json-logger>=2.0",
    "rich>=13.0",
]
config = [
    "pyyaml>=6.0",
]

[project.scripts]
magellon-sdk = "magellon_sdk.cli:main"

[project.urls]
Homepage = "https://magellon.org/magellonsdk"
Repository = "https://github.com/magellon-org/magellon"
Issues = "https://github.com/magellon-org/magellon/issues"
